#134 refactor: queries 파일명 컨벤션 변경
