+++ src/hooks/mutations/useUpdateTodo.tsx
@@ -1,7 +1,7 @@
 import { updateTodo } from '@/apis';
 import { MutateOptionsType, UpdateTodoParamType } from '@/hooks/mutations/types';
+import { todoKeys } from '@/hooks/queries';
 import { InfiniteQueryData } from '@/hooks/queries/types';
-import { todoKeys } from '@/hooks/queries/useTodo';
 import { ApiResponse, CreateAndUpdateResponseType, ToDoDetail } from '@/types';
 import { optimisticUpdateMutateHandler } from '@/utils/optimisticUpdateMutateHandler';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
